% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFrog2FeatureDataHandler.R
\docType{data}
\name{cFrog2FeatureDataHandler}
\alias{cFrog2FeatureDataHandler}
\title{cFrogdataHandler}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
cFrog2FeatureDataHandler
}
\arguments{
\item{dtfrog}{}

\item{objFeatures}{}

\item{...}{}

\item{document.labels}{}

\item{document.labels.columns}{}

\item{do.return}{}

\item{do.save.internally}{}
}
\description{
cFrogdataHandler
}
\examples{
Title

}
\keyword{datasets}
