% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFrogDataTransformator.R
\docType{data}
\name{cFrogDataTransformator}
\alias{cFrogDataTransformator}
\title{cFrogDataSelector}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
cFrogDataTransformator
}
\description{
cFrogDataSelector
}
\keyword{datasets}
