% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_evaluation.R,
%   R/functionsFeatureEvaluation.R
\name{feature_to_t_values}
\alias{feature_to_t_values}
\alias{feature_to_t_values}
\title{Calculate test statistics for predictive accuracy of a feature.}
\usage{
feature_to_t_values(dtable, dependents, feature, controls = NULL,
  family = binomial(link = "logit"))

feature_to_t_values(dtable, dependents, feature, controls = NULL,
  family = binomial(link = "logit"))
}
\arguments{
\item{dtable}{A data.table object with dependents and features.}

\item{dependents}{A vector denoting the names of the dependent variable(s).}

\item{feature}{A vector denoting the name of the feature colomn.}

\item{controls}{A character vector denoting the name of the control variables; if specified, tests statistics provide information on the incremental relevance of a feature given the controls.}

\item{dtable}{A data.table object with dependents and features.}

\item{dependents}{A vector denoting the names of the dependent variable(s).}

\item{feature}{A vector denoting the name of the feature colomn.}

\item{controls}{A character vector denoting the name of the control variables; if specified, tests statistics provide information on the incremental relevance of a feature given the controls.}
}
\value{
a modified data.table object with strings instead of factors.

a modified data.table object with strings instead of factors.
}
\description{
Calculate test statistics for predictive accuracy of a feature.

Calculate test statistics for predictive accuracy of a feature.
}
