% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsFrogDataHandling.R
\name{filter_frog}
\alias{filter_frog}
\title{Filter a frog object}
\usage{
filter_frog(dtable, exclude)
}
\arguments{
\item{exclude}{a named list of vectors with elements to exclude; valid elements are 'docids', 'majorpos', 'words' and 'lemmas'.}

\item{dtfrog}{a frog data.table object}
}
\value{
A filtered frog data.table object
}
\description{
Filter a frog object
}
\examples{
\dontrun{
dtfrog.filtered <- filter_frog(dtfrog, exclude = list(majorpos=c("LET")))
}
}
