% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsConvenienceR.R
\name{FixStringsFactors}
\alias{FixStringsFactors}
\title{Convert all factor variables in a data.table to st44rings.}
\usage{
FixStringsFactors(dtable)
}
\arguments{
\item{dtable}{the imput data.table}
}
\value{
a modified data.table object with strings instead of factors.
}
\description{
Convert all factor variables in a data.table to st44rings.
}
\examples{
\donttest{ 
x.fixed <- x \%>\% FixStringsFactors 
}
}
