% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsConvenienceR.R
\name{\%frogmerge\%}
\alias{\%frogmerge\%}
\title{Merge two frog data.table objects}
\usage{
a \%frogmerge\% b
}
\arguments{
\item{x}{A frog data.table}

\item{y}{A frog data.tables}
}
\value{
The union of \code{x} and \code{y}
}
\description{
Merge two frog data.table objects
}
\examples{
\dontrun{
c(1,2) \%union\% c(3,4)
x \%union\% y \%union\% z
}
}
