% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsConvenienceR.R
\name{\%if\%}
\alias{\%if\%}
\title{Alternative if}
\usage{
x \%if\% y
}
\arguments{
\item{x}{Command to be executed; must be parenthesized.}
}
\value{
y The condition to satisfy; must be parenthesizd.
}
\description{
Alternative if
}
\examples{
\dontrun{
(print(test)) \%if\% (1 == 1) 
}
}
