% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsConvenienceR.R
\name{\%intersect\%}
\alias{\%intersect\%}
\title{Performs set intersect}
\usage{
x \%intersect\% y
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
The union of \code{x} and \code{y}
}
\description{
Performs set intersect
}
\examples{
\dontrun{
c(1,2) \%union\% c(3,4)
x \%union\% y \%union\% z
}
}
