% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsConvenienceR.R
\name{\%+\%}
\alias{\%+\%}
\title{Concatenate strings.}
\usage{
x \%+\% y
}
\arguments{
\item{x, y}{A string or a vector of strings; numerics are converted to strings.}
}
\value{
The concatenated strings. Elements are recycled.
}
\description{
Concatenate strings.
}
\examples{
\dontrun{
c("a","b") \%+\% c("d","e")
} 
}
