% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frogdatachecks.R, R/functionsFrogDatachecks.R
\name{is.frog}
\alias{is.frog}
\alias{is.frog}
\title{Check whether the object is valid frog data.table.}
\usage{
is.frog(dtfrog, require_all_original_columns = TRUE, require_docid = TRUE,
  require_sent = TRUE, require_position = TRUE, require_word = FALSE,
  require_lemma = FALSE, require_morph = FALSE, require_pos = FALSE,
  require_prob = FALSE, require_ner = FALSE, require_chunk = FALSE,
  require_parse1 = FALSE, require_parse2 = FALSE, require_majorpos = TRUE,
  require_unique_keys = TRUE, allow_extra_columns = FALSE)

is.frog(dtfrog, require_all_original_columns = TRUE, require_docid = TRUE,
  require_sent = TRUE, require_position = TRUE, require_word = FALSE,
  require_lemma = FALSE, require_morph = FALSE, require_pos = FALSE,
  require_prob = FALSE, require_ner = FALSE, require_chunk = FALSE,
  require_parse1 = FALSE, require_parse2 = FALSE, require_majorpos = TRUE,
  require_unique_keys = TRUE, allow_extra_columns = FALSE)
}
\arguments{
\item{dtfrog}{a suspected frog data.table object}

\item{exclude}{a named list of vectors with elements to exclude; valid elements are 'docids', 'majorpos', 'words' and 'lemmas'.}

\item{dtfrog}{a suspected frog data.table object}

\item{exclude}{a named list of vectors with elements to exclude; valid elements are 'docids', 'majorpos', 'words' and 'lemmas'.}
}
\value{
A filtered frog data.table object

A filtered frog data.table object
}
\description{
Check whether the object is valid frog data.table.

Check whether the object is valid frog data.table.
}
\examples{
\dontrun{
dtfrog.filtered <- filter_frog(dtfrog, exclude = list(majorpos=c(\\"LET\\")))
}
\dontrun{
dtfrog.filtered <- filter_frog(dtfrog, exclude = list(majorpos=c(\\"LET\\")))
}
}
