% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsFrogDataHandling.R
\name{select_frog}
\alias{select_frog}
\title{Select items from a frog object}
\usage{
select_frog(dtfrog, select = list(docids = NULL, majorpos = NULL, words =
  NULL, lemmas = NULL, logic = "OR"))
}
\arguments{
\item{dtfrog}{a frog data.table object}

\item{select}{a named list of vectors with elements to exclude; valid elements are 'docids', 'majorpos', 'words' and 'lemmas'.}
}
\value{
A filtered frog data.table object
}
\description{
Select items from a frog object
}
\examples{
\dontrun{
dtfrog.filtered <- filter_frog(dtfrog, exclude = list(majorpos=c("LET")))
}
}
